User stats should have its own generator.
The 4.2.0 release already uses the cached_stats generator, so we should not add to it lest we break/complicate the upgrade path.  This takes the modifications to that generator and breaks them out into their own generator, which will enable adopters to upgrade cleanly from 4.2 to 4.3.
